Refactor deeplink
