+++ app/deeplink/methods/index.js
@@ -1,0 +1,4 @@
+export { default as fulfillauction } from './fulfillauction';
+export { default as openmanager } from './openmanager';
+export { default as openname } from './openname';
+export { default as updaterecord } from './updaterecord';
