+++ app/deeplink/methods/openmanager.js
@@ -1,0 +1,7 @@
+import { history } from '../../store/configureStore';
+
+export default message => {
+  const name = new URL(message).searchParams.get('name');
+
+  history.push(`/domain_manager/${name}`);
+};
